\name{exampledata}
\alias{exampledata}
\docType{data}
\title{
The example data set
}
\description{
This is an example data set used only to fit the qardl model
}
\usage{data("exampledata")}
\format{
  A data frame with 10000 observations on the following 4 variables.
  \describe{
    \item{\code{...1}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{z1}}{a numeric vector}
    \item{\code{z2}}{a numeric vector}
  }
}

\examples{
data(exampledata)
## maybe str(exampledata) ; plot(exampledata) ...
}
\keyword{datasets}
