% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyptest.R
\name{hyptest}
\alias{hyptest}
\title{hyptest function}
\usage{
hyptest(formula, data, maxlag = 7, tau = NULL)
}
\arguments{
\item{formula}{y~z1+z2}

\item{data}{the dataframe}

\item{maxlag}{maximum lag number}

\item{tau}{the quantile(s) to be estimated, this is generally a number strictly between 0 and 1}
}
\value{
the short-run phi and gamma wald test and the long-run beata wald test
}
\description{
hyptest function
}
\examples{

# Quantile ARDL regression
# load data
data(exampledata)
# Fit the model
hyp=hyptest(y~z1+z2,exampledata,maxlag=7, tau=c(0.2,0.5,0.75))
summary(hyp)

}
