% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lagm}
\alias{lagm}
\title{Function lagm: lagged matrix}
\usage{
lagm(m, nLags)
}
\arguments{
\item{nLags}{the number of lags

#load data
data(exampledata)
m=as.matrix(exampledata$y)
lagm(m,2)}

\item{x}{is a matrix}
}
\description{
Function lagm: lagged matrix
}
\keyword{internal}
