% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lags}
\alias{lags}
\title{Function lags which return a lagged matrix}
\usage{
lags(x, k = 1)
}
\arguments{
\item{x}{is a matrix}

\item{k}{the number of lags}
}
\value{
A lagged matrix
}
\description{
Function lags which return a lagged matrix
}
\examples{

 data(exampledata)
lags(exampledata$y,2)

}
\keyword{internal}
