% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{matsplitter}
\alias{matsplitter}
\title{Function matsplitter}
\usage{
matsplitter(M, r, c)
}
\arguments{
\item{M}{matrix}

\item{r}{matrix rows}

\item{c}{matrix columns

aaa=matrix(c(seq(50)),25,2)
matsplitter(aaa,5,2)}
}
\description{
Function matsplitter
}
\keyword{internal}
