% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{mlags}
\alias{mlags}
\title{Function mlags which return a lagged matrix}
\usage{
mlags(x, nlags)
}
\arguments{
\item{x}{is a matrix}

\item{nlags}{the number of lags}
}
\value{
A lagged matrix
}
\description{
Function mlags which return a lagged matrix
}
\examples{

 data(exampledata)
 mlags(as.matrix(exampledata$y),2)

}
\keyword{internal}
