% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{nlagm}
\alias{nlagm}
\title{Function nlagm return a  lagged matrix}
\usage{
nlagm(x, nlags)
}
\arguments{
\item{x}{is a matrix}

\item{nlags}{the number of lags}
}
\value{
A lagged matrix
}
\description{
Function nlagm return a  lagged matrix
}
\examples{

 data(exampledata)
 nlagm(as.matrix(exampledata$y),2)

}
\keyword{internal}
