% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{nlagm1}
\alias{nlagm1}
\title{Function nlagm1 which return a lagged matrix}
\usage{
nlagm1(x, nlags)
}
\arguments{
\item{x}{is a matrix}

\item{nlags}{the number of lags}
}
\value{
A lagged matrix
}
\description{
Function nlagm1 which return a lagged matrix
}
\examples{

data(exampledata)
nlagm1(as.matrix(exampledata$y),2)

}
\keyword{internal}
