% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{qardl}
\alias{qardl}
\title{Qardl function}
\usage{
qardl(formula, data, maxlag = 4, tau = NULL)
}
\arguments{
\item{formula}{y~z1+z2}

\item{data}{the dataframe}

\item{maxlag}{maximum lag number}

\item{tau}{the quantile(s) to be estimated, this is generally a number strictly between 0 and 1}
}
\value{
the short-run and the long-run estimated coefficients of the QARDL model
}
\description{
Qardl function
}
\examples{

# Quantile ARDL regression
# load data
data(exampledata)
# Fit the model
reg=qardl(y~z1+z2,exampledata,maxlag=7, tau=0.5)
reg

}
