% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hyptest.R
\name{summary.hyptest}
\alias{summary.hyptest}
\title{Summary of a hyptest}
\usage{
\method{summary}{hyptest}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\value{
an object of the S3 class \code{summary.hyptest} with the
following components:
}
\description{
\code{summary} method for a \code{\link{hyptest}}.
}
