% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ardl.R
\name{summary.qardl}
\alias{summary.qardl}
\title{Summary of a ardl model}
\usage{
\method{summary}{qardl}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\value{
an object of the S3 class \code{summary.qardl} with the
following components:
}
\description{
\code{summary} method for a \code{\link{qardl}} model.
}
