% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{wtestlr}
\alias{wtestlr}
\title{Function wtelslr beta}
\usage{
wtestlr(cof, covar, rca, rsm, data)
}
\arguments{
\item{cof}{the estimated long-run parameter by qardl}

\item{covar}{estimated covariance matrix}

\item{rca}{R matrix in the null hypothesis}

\item{rsm}{r matrix in the null hypothesis}

\item{data}{the same data set used for qardl

#'data(exampledata)
hyp=hyptest(y~z1+z2,exampledata,maxlag=7,tau=c(0.25,0.5,0.75))
wlr=wtestlr(hyp$bigbt,hyp$lrvcov3, hyp$ca1,hyp$sm1,exampledata)
wlr}
}
\description{
Function wtelslr beta
}
\keyword{internal}
