% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{wtestlsr}
\alias{wtestlsr}
\title{Function wtelslsr gamma and phi}
\usage{
wtestlsr(cof, covar, rca, rsm, data)
}
\arguments{
\item{cof}{the estimated short-run parameter by qardl}

\item{covar}{estimated covariance matrix}

\item{rca}{R matrix in the null hypothesis}

\item{rsm}{r matrix in the null hypothesis}

\item{data}{the same data set used for qardl

data(exampledata)
hyp=hyptest(y~z1+z2,exampledata,maxlag=7,tau=c(0.25,0.5,0.75))
wsrg=wtestlsr(hyp$bigdam, hyp$bigff, hyp$ca1, hyp$sm1, exampledata)
wsrg}
}
\description{
Function wtelslsr gamma and phi
}
\keyword{internal}
