% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lagm}
\alias{lagm}
\title{Function lagm which return a lagged matrix}
\usage{
lagm(m, nLags)
}
\arguments{
\item{m}{is a matrix}

\item{nLags}{the number of lags}
}
\value{
A lagged matrix
}
\description{
Function lagm which return a lagged matrix
}
\examples{

 data(exampledata)
 lagm(as.matrix(exampledata$y),2)
}
\keyword{internal}
