% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{mlags}
\alias{mlags}
\title{Function mlags: lagged matrix}
\usage{
mlags(x, nlags)
}
\arguments{
\item{x}{is a matrix}

\item{nlags}{the number of lags

#load data
data(exampledata)
m=as.matrix(exampledata$y)
mlags(m,2)}
}
\description{
Function mlags: lagged matrix
}
\keyword{internal}
