% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{nlagm}
\alias{nlagm}
\title{Function nlagm: lagged matrix}
\usage{
nlagm(x, nlags)
}
\arguments{
\item{x}{is a matrix}

\item{nlags}{the number of lags

#load data
data(exampledata)
m=as.matrix(exampledata$y)
nlagm(m,2)}
}
\description{
Function nlagm: lagged matrix
}
\keyword{internal}
