% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{nlagm1}
\alias{nlagm1}
\title{Function nlagm1: lagged matrix}
\usage{
nlagm1(x, nlags)
}
\arguments{
\item{x}{is a matrix}

\item{nlags}{the number of lags

#load data
data(exampledata)
m=as.matrix(exampledata$y)
nlagm1(m,2)}
}
\description{
Function nlagm1: lagged matrix
}
\keyword{internal}
